x86/DtsTerminal/requirements.txt
pyserial
pycryptoplus
aioconsole